git convert

+++ src/recreatedirectory.py
@@ -24,10 +24,16 @@ def recreate_directory(pig_root: Path, commit_hash: str) -> None:
         if not dest_path.parent.exists():
             dest_path.parent.mkdir(parents=True)
         compressed_file_path: Path  = pig_root / ".pig" / "compressed-files" / fileinfo.hash
-        with gzip.open(compressed_file_path, "rb") as f_in:
-            with open(dest_path, "wb") as f_out:
-                shutil.copyfileobj(f_in, f_out)
+        # print(f"Recreating file {filepath}...")
+        try:
+            with gzip.open(compressed_file_path, "rb") as f_in:
+                with open(dest_path, "wb") as f_out:
+                    shutil.copyfileobj(f_in, f_out)
+        except Exception as e:
+            print(f"Error recreating file {filepath}: {e}")
     clear_directory(pig_root, ignoreFiles={".pig"})
+    if not tmp_dir.exists():
+        return
     for item in tmp_dir.iterdir():
         shutil.move(str(item), str(pig_root))
     clear_directory(tmp_dir)
